fix import for ball_tree and kd_tree
